Adds some general load code to base object
